requirements.txt
autopep8==1.6.0
cachelib==0.5.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cloudinary==1.29.0
cs50==8.0.1
distlib==0.3.4
filelock==3.4.2
flake8==4.0.1
Flask==2.0.2
Flask-Session==0.4.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.22.1
Pillow==9.0.0
platformdirs==2.4.1
pycodestyle==2.8.0
pyflakes==2.4.0
pygame==2.1.0
requests==2.26.0
scipy==1.7.3
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
termcolor==1.1.0
toml==0.10.2
urllib3==1.26.7
virtualenv==20.11.0
Werkzeug==2.0.2
